[AC-7953] De-specialize assert_ui_notification function
