+++ web/impact/impact/tests/test_cancel_office_hour_reservation.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 from accelerator.tests.factories import (
     MentorProgramOfficeHourFactory,
 )
-from ..tests.api_test_case import APITestCase
+from .office_hours_test_case import OfficeHoursTestCase
 from ..v1.views import (
     CancelOfficeHourReservationView,
     formatted_success_notification,
@@ -18,7 +18,7 @@ from ..v1.views.cancel_office_hour_reservation_view import (
 from ..v1.views.utils import localized_office_hour_start_time
 
 
-class TestCancelOfficeHourReservationView(APITestCase):
+class TestCancelOfficeHourReservationView(OfficeHoursTestCase):
     fail_header = FAIL_HEADER
     success_header = SUCCESS_HEADER
